Slight changes to ver2
Tested the baseline model with a learning rate of 0.001 and batch size of 128 for 15 epochs. Results are shown in Baseline_model_accuracy image. Rest of setup was not changed except for adding a code block for finetuning the model after freezing and then unfreezing the backbone. Training appeared to be much faster in this case with better results generated.